+++ apiary-metastore-events/kafka-metastore-events/kafka-metastore-integration-tests/src/test/java/com/expediagroup/apiary/extensions/events/metastore/kafka/messaging/KafkaApiaryMetastoreIntegrationTest.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (C) 2018-2019 Expedia, Inc.
+ * Copyright (C) 2018-2020 Expedia, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,13 +19,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
+import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaConsumerProperty.CLIENT_ID;
 import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaConsumerProperty.GROUP_ID;
 import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaIntegrationTestUtils.buildPartition;
 import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaIntegrationTestUtils.buildQualifiedTableName;
 import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaIntegrationTestUtils.buildTable;
 import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaIntegrationTestUtils.buildTableParameters;
 import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaProducerProperty.BOOTSTRAP_SERVERS;
-import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaProducerProperty.TOPIC;
+import static com.expediagroup.apiary.extensions.events.metastore.kafka.messaging.KafkaProducerProperty.TOPIC_NAME;
 
 import java.util.ArrayList;
 import java.util.Properties;
@@ -74,7 +75,6 @@ public class KafkaApiaryMetastoreIntegrationTest {
   public static final SharedKafkaTestResource KAFKA = new SharedKafkaTestResource();
 
   private static final long TEST_TIMEOUT_MS = 10000;
-  private static final String TOPIC_NAME = "topic";
   private static Configuration CONF = new Configuration();
 
   private static KafkaMetaStoreEventListener kafkaMetaStoreEventListener;
@@ -86,8 +86,9 @@ public class KafkaApiaryMetastoreIntegrationTest {
   public static void init() {
     CONF.set(BOOTSTRAP_SERVERS.key(), KAFKA.getKafkaConnectString());
     CONF.set(GROUP_ID.key(), "1");
-    CONF.set(TOPIC.key(), TOPIC_NAME);
-    KAFKA.getKafkaTestUtils().createTopic(TOPIC_NAME, 1, (short) 1);
+    CONF.set(CLIENT_ID.key(), "client");
+    CONF.set(TOPIC_NAME.key(), "topic");
+    KAFKA.getKafkaTestUtils().createTopic("topic", 1, (short) 1);
 
     kafkaMetaStoreEventListener = new KafkaMetaStoreEventListener(CONF);