+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 5.0.3 - TBD
+### Changed
+- `client.id` is now required to configure Kafka Receiver and Listener.
+- `topic` Kafka configuration parameter is now called `topic.name`.
+- No longer sending MetaStore uris with Kafka message.
+
 ## 5.0.2 - 2020-02-04
 ### Changed
 - `eg-oss-parent` version updated to 1.2.0 (was 1.1.0).